Moved all attributes into a separate dir;
