+++ animals/__init__.py
@@ -1,12 +1,4 @@
-from .identifiable import Identifiable
 from .animal import Animal
-from .terrestrial import Terrestrial
-from .aquatic import Aquatic
-from .freshwater import Freshwater
-from .saltwater import Saltwater
-from .flying import Flying
-from .walking import Walking
-from .swimming import Swimming
 from .river_dolphin import RiverDolphin
 from .gold_dust_day_gecko import GoldDustDayGecko
 from .hawaiian_spider import HawaiianHappyFaceSpider
