+++ animals/attributes/__init__.py
@@ -1,0 +1,8 @@
+from .identifiable import Identifiable
+from .terrestrial import Terrestrial
+from .aquatic import Aquatic
+from .freshwater import Freshwater
+from .saltwater import Saltwater
+from .flying import Flying
+from .walking import Walking
+from .swimming import Swimming
