+++ animals/attributes/walking.py
