+++ environments/river.py
@@ -1,4 +1,3 @@
-from animals import Aquatic
 from environments import Environment
 from animals import RiverDolphin
 
@@ -6,19 +5,20 @@ from animals import RiverDolphin
 class River(Environment):
 
     def __init__(self, name):
-      Environment.__init__(self, name)
-      
+        Environment.__init__(self, name)
 
     def add_animal(self, animal):
         try:
             if animal.aquatic and animal.cell_type == "hypertonic":
                 self.animals.append(animal)
         except AttributeError:
-            raise AttributeError("Cannot add non-aquatic, or saltwater animals to a river")
+            raise AttributeError(
+                "Cannot add non-aquatic, or saltwater animals to a river")
 
     def add_plant(self, plant):
         try:
             if plant.freshwater and plant.requires_current:
                 self.plants.append(plant)
         except AttributeError:
-            raise AttributeError("Cannot add plants that require brackish water or stagnant water to a river biome")
+            raise AttributeError(
+                "Cannot add plants that require brackish water or stagnant water to a river biome")
